+++ tests/Extensions/Abstractions.UnitTests/Accessors/AccessorTests.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
+using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.Omex.Extensions.Abstractions;
 using Microsoft.Omex.Extensions.Abstractions.Accessors;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -15,7 +16,7 @@ namespace Microsoft.Omex.Extensions.Hosting.Services.UnitTests
 		public void ValuePassedInConstructor_ProperlyHandlesActions()
 		{
 			object value = new object();
-			Accessor<object> accessor = new Accessor<object>(value);
+			Accessor<object> accessor = new Accessor<object>(new NullLogger<Accessor<object>>(), value);
 			IAccessor<object> publicAccessor = accessor;
 
 			Assert.AreEqual(value, publicAccessor.Value);
@@ -29,7 +30,7 @@ namespace Microsoft.Omex.Extensions.Hosting.Services.UnitTests
 		public void ValueAfterInitialization_ProperlyHandlesActions()
 		{
 			object value = new object();
-			Accessor<object> accessor = new Accessor<object>();
+			Accessor<object> accessor = new Accessor<object>(new NullLogger<Accessor<object>>());
 			IAccessor<object> publicAccessor = accessor;
 			IAccessorSetter<object> setter = accessor;
