+++ tests/Extensions/Hosting.Services.UnitTests/OmexServiceFabricContextTests.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license.
 
 using System.Fabric;
+using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.Omex.Extensions.Abstractions.Accessors;
 using Microsoft.Omex.Extensions.Logging;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -23,7 +24,7 @@ namespace Microsoft.Omex.Extensions.Hosting.Services.UnitTests
 		public void TestContextSet<TContext>(TContext serviceContext)
 			where TContext : ServiceContext
 		{
-			Accessor<TContext> accessor = new Accessor<TContext>();
+			Accessor<TContext> accessor = new Accessor<TContext>(new NullLogger<Accessor<TContext>>());
 			IAccessorSetter<TContext> setter = accessor;
 			IServiceContext context = new OmexServiceFabricContext(accessor);
